requirements.txt
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
et-xmlfile==1.1.0
h11==0.14.0
idna==3.6
MouseInfo==0.1.3
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.1.4
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
XlsxWriter==3.1.9
